requirements.txt
duckdb==1.1.3
openai-whisper
jiwer
bitsandbytes
transformers
peft
accelerate
scipy
soundfile
librosa
pandas
tabulate
sentencepiece
torchaudio
edge-tts
huggingface-hub
nest_asyncio
pyarrow
